KAFKA-7986: Distinguish logging from different ZooKeeperClient instances (#6493)
A broken can have more than one instance of ZooKeeperClient. For example, SimpleAclAuthorizer creates a separate ZooKeeperClient instance when configured.

This commit makes it possible to optionally specify the name for the ZooKeeperClient instance. The name is specified only for a broker's ZooKeeperClient instances, but not for commands' and tests'.

Reviewers: Jun Rao <junrao@gmail.com>