MINOR: Move log statement stating producer closed after possible exception (#6606)
In the KafkaProducer#close method we have a debug log statement Kafka producer has been closed then a few lines later a KafkaException can occur.
This could be confusing to users, so this PR simply moves the log statement to after the possible exception to avoid confusing information in the logs.

Reviewers:  Guozhang Wang <wangguoz@gmail.com>