KAFKA-9098: When users name repartition topic, use the name for the repartition filter, source and sink node. (#7598)
When users specify a name for a repartition topic, we should use the same name for the repartition filter, source, and sink nodes. With the addition of KIP-307 if users go to the effort of naming every node in the topology having processor nodes with generated names is inconsistent behavior.

Updated tests in the streams test suite.

Reviewers: John Roesler <john@confluent.io>, Christopher Pettitt <cpettitt@confluent.io>