MINOR. Replace Utils::readFileAsString method to read file as stream (#7208)
The current Utils::readFileAsString method creates a FileChannel and
memory maps file and copies its content to a String and returns it. But
that means that we need to know the size of the file in advance. This
precludes us from reading files whose size is not known in advance, i.e.
any file opened with flag S_IFIFO.

This change updates the method to use stream to read the content of the file.
It has couple of practical advantages:

Allows bash process substitution to pass in strings as file. So we can
say `./bin/kafka-reassign-partitions.sh --reassignment-json-file <(echo
"reassignment json")

When adding systest for commands that take file, we don't have to
create real physical file. Instead we can just dump the content of the
file on the command line.

Reviewers: Ismael Juma <ismael@juma.me.uk>