MINOR: Only start log dir fetcher after LeaderAndIsr epoch validation (#8460)
Currently a `LeaderAndIsr` request with a stale leader epoch for some partition may still result in the starting of the log dir fetcher for that partition (if the future log exists). I am not sure if this causes any correctness problem since we don't use any state from the request to start the fetcher, but it seems unnecessary to rely on this side effect.

Reviewers: Jun Rao <junrao@gmail.com>