KAFKA-10371; Partition reassignments can result in crashed ReplicaFetcherThreads (#9140)
The patch https://github.com/apache/kafka/pull/8672 introduced a bug leading to crashing the replica fetcher threads. The issue is that https://github.com/apache/kafka/pull/8672 deletes the Partitions prior to stopping the replica fetchers. As the replica fetchers relies access the Partition in the ReplicaManager, they crash with a NotLeaderOrFollowerException that is not handled.

This PR reverts the code to the original ordering to avoid this issue.

The regression was caught and validated by our system test: `kafkatest.tests.core.reassign_partitions_test`.

Reviewers: Vikas Singh <vikas@confluent.io>, Jason Gustafson <jason@confluent.io>