MINOR: fix some bugs in ControllerApis.scala (#10505)
Fix some cases where ControllerApis was blocking on the controller
thread.  This should not be necessary, since the controller thread can
just interface directly with the network threads.

Fix some cases where ControllerApis wasn't doing authorization
correctly.  Since the previous release of KRaft did not support
authorizers, this bug is not as severe as it could have been, but it
still needs to be fixed.  Add authorization unit tests for each API.

Add support for the deprecated ALTER_CONFIGS API in ControllerApis.  It
was already supported in QuorumController, but wasn't exposed
previously.

Fix how we validate duplicate config resources and unknown config
resource types in ControllerApis.  Duplicates should yield an
INVALID_REQUEST error, and unknown resource types should give an error
with the corresponding numerical resource type and UNSUPPORTED_VERSION.
Fix some redaction code in RequestChannel that was throwing an exception
when duplicate config resources were present in the request.

Fix a comment in ControllerApis#deleteTopics that no longer reflects
what the code is doing when we don't have "describe" permission.

Add function stubs for the KIP-455 reassignment APIs in ControllerApis
and QuorumController.

Reviewers: Chia-Ping Tsai <chia7712@gmail.com>, David Arthur <mumrah@gmail.com>