KAFKA-8529; Flakey test ConsumerBounceTest#testCloseDuringRebalance (#11097)
When the replica fetcher receives a top-level error in the fetch response, it marks all partitions are failed and adds a backoff delay before resuming fetching from them. In addition to this, there is an additional backoff enforced after the top-level error is handled, so we end up waiting twice the configured backoff time before resuming. This patch removes this extra backoff.

Reviewers: Jason Gustafson <jason@confluent.io>