KAFKA-13319: Do not commit empty offsets on producer (#11362)
We observed on the broker side that txn-offset-commit request with empty topics are received. After checking the source code I found there's on place on Streams which is unnecessarily sending empty offsets. This PR cleans up the streams layer logic a bit to not send empty offsets, and at the same time also guard against empty offsets at the producer layer as well.

Reviewers: Luke Chen <showuon@gmail.com>, Jason Gustafson <jason@confluent.io>